requirements.txt
opencv-python
pillow
reportlab
easyocr
psutil>=5.9.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
pdf2image>=1.16.0
numpy>=1.24.0
scikit-image>=0.21.0
easyocr>=1.7.0
